backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pillow==10.2.0
pydantic==2.6.4
numpy>=1.24
torch>=2.1
torchvision>=0.16
